more tolerant reading data with missing attributes
when changing your schema of object. eg. adding a field/attribute, then the knockoutFire function exits with error, because the "map" cannot be matched with the data coming from the database. 
with this   

if( typeof self() == 'null' || options.map == true ) return; 

it will still work, and one can read and write data.